De-deify settings and terminal modules
