+++ desktop/app-settings/keys.cts
@@ -1,0 +1,18 @@
+export const gitCommitMessageModelKey = "gitCommitMessageModel";
+export const gitCommitMessageThinkingLevelKey = "gitCommitMessageThinkingLevel";
+export const skillCreatorModelKey = "skillCreatorModel";
+export const skillCreatorThinkingLevelKey = "skillCreatorThinkingLevel";
+export const composerStreamingBehaviorKey = "composerStreamingBehavior";
+export const dictationModelIdKey = "dictationModelId";
+export const dictationMaxDurationSecondsKey = "dictationMaxDurationSeconds";
+export const showDictationButtonKey = "showDictationButton";
+export const favoriteFoldersKey = "favoriteFolders";
+export const projectImportStateKey = "projectImportState";
+export const preferredProjectLocationKey = "preferredProjectLocation";
+export const initializeGitOnProjectCreateKey = "initializeGitOnProjectCreate";
+export const gitOpsDefaultModeKey = "gitOpsDefaultMode";
+export const gitDiffBaselineDefaultKey = "gitDiffBaselineDefault";
+export const gitDiffRenderModeDefaultKey = "gitDiffRenderModeDefault";
+export const projectDeletionModeKey = "projectDeletionMode";
+export const useAgentsSkillsPathsKey = "useAgentsSkillsPaths";
+export const piTuiTakeoverKey = "piTuiTakeover";
