+++ src/app/components/workspace/terminal/TerminalViewport.tsx
@@ -1,19 +1,41 @@
 import { Terminal, type TerminalHandle, type WTerm } from "@wterm/react";
-import { type CSSProperties, useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { getPersistedSessionPath } from "../../../../../shared/session-paths";
 import type { TerminalEvent } from "../../../desktop/types";
 import {
   closeDesktopTerminal,
-  getDesktopTerminalStatus,
   openDesktopTerminal,
   resizeDesktopTerminal,
-  statDesktopTerminalSessionFile,
   subscribeDesktopTerminal,
   writeDesktopTerminal,
 } from "../../../hooks/useDesktopTerminal";
 import { cn } from "../../../utils/cn";
-
-type TerminalBackgroundCssVar = "--terminal-bg" | "--workspace" | "--sidebar";
+import {
+  cancelScheduledTerminalClose,
+  scheduleTerminalClose,
+  scheduleTerminalCloseAfterSessionFileIdle,
+} from "./terminalViewportSessionLifecycle";
+import {
+  DEFAULT_MAX_KEEP_ALIVE_MS_ON_UNMOUNT,
+  DEFAULT_TERMINAL_COLS,
+  DEFAULT_TERMINAL_ROWS,
+  MAX_PENDING_TERMINAL_EVENTS,
+  MIN_INITIAL_TERMINAL_COLS,
+  MIN_INITIAL_TERMINAL_ROWS,
+  clampTerminalHistory,
+  clearTerminal,
+  findTerminalLinkAtPoint,
+  hasSelectionInside,
+  hasVisibleTerminalHistory,
+  isTerminalElementNearBottom,
+  isUsableTerminalSize,
+  measureTerminalSize,
+  normalizeTerminalDimension,
+  terminalStyleVars,
+  terminalWrapperStyle,
+  type TerminalBackgroundCssVar,
+  writeSystemMessage,
+} from "./terminalViewportUtils";
 
 type TerminalViewportProps = {
   projectId: string;
@@ -27,450 +49,6 @@ type TerminalViewportProps = {
   backgroundCssVar?: TerminalBackgroundCssVar;
   className?: string;
 };
-
-type TerminalLinkMatch = {
-  text: string;
-  start: number;
-  end: number;
-};
-
-const pendingTerminalCloseTimers = new Map<string, ReturnType<typeof setTimeout>>();
-const pendingTerminalCloseGenerations = new Map<string, number>();
-
-const CLEAR_TERMINAL_SEQUENCE = "\u001b[2J\u001b[3J\u001b[H";
-const MAX_PENDING_TERMINAL_EVENTS = 200;
-const MAX_FRONTEND_HISTORY_CHARS = 200_000;
-const TRIMMED_FRONTEND_HISTORY_CHARS = 160_000;
-const TERMINAL_LINK_PATTERN = /https?:\/\/[^\s)\]}]+/g;
-const DEFAULT_TERMINAL_COLS = 80;
-const DEFAULT_TERMINAL_ROWS = 24;
-const MIN_INITIAL_TERMINAL_COLS = 20;
-const MIN_INITIAL_TERMINAL_ROWS = 5;
-const MIN_USABLE_TERMINAL_COLS = 2;
-const MIN_USABLE_TERMINAL_ROWS = 2;
-const TERMINAL_STICKY_BOTTOM_THRESHOLD_PX = 24;
-const DEFAULT_MAX_KEEP_ALIVE_MS_ON_UNMOUNT = 12 * 60 * 60 * 1_000;
-const MAX_TERMINAL_STATUS_FAILURES_BEFORE_CLOSE = 2;
-const ANSI_ESCAPE = String.fromCharCode(27);
-type SessionFileStat = { mtimeMs: number; size: number };
-
-function hasVisibleTerminalHistory(history: string) {
-  return (
-    history
-      .split(ANSI_ESCAPE)
-      .map((segment, index) =>
-        index === 0 ? segment : segment.replace(/^\[[0-?]*[ -/]*[@-~]/, ""),
-      )
-      .join("")
-      .trim().length > 0
-  );
-}
-
-function closeScheduledTerminal(sessionId: string, generation: number) {
-  if (pendingTerminalCloseGenerations.get(sessionId) !== generation) {
-    return;
-  }
-
-  pendingTerminalCloseTimers.delete(sessionId);
-  void closeDesktopTerminal({ sessionId }).finally(() => {
-    if (pendingTerminalCloseGenerations.get(sessionId) === generation) {
-      pendingTerminalCloseGenerations.delete(sessionId);
-    }
-  });
-}
-
-function cancelScheduledTerminalClose(sessionId: string) {
-  const timer = pendingTerminalCloseTimers.get(sessionId);
-  if (!timer) {
-    pendingTerminalCloseGenerations.delete(sessionId);
-    return;
-  }
-
-  pendingTerminalCloseGenerations.set(
-    sessionId,
-    (pendingTerminalCloseGenerations.get(sessionId) ?? 0) + 1,
-  );
-  clearTimeout(timer);
-  pendingTerminalCloseTimers.delete(sessionId);
-}
-
-function scheduleTerminalClose(sessionId: string, delayMs: number) {
-  cancelScheduledTerminalClose(sessionId);
-  const generation = (pendingTerminalCloseGenerations.get(sessionId) ?? 0) + 1;
-  pendingTerminalCloseGenerations.set(sessionId, generation);
-
-  const timer = setTimeout(() => {
-    if (pendingTerminalCloseGenerations.get(sessionId) !== generation) {
-      return;
-    }
-
-    closeScheduledTerminal(sessionId, generation);
-  }, delayMs);
-
-  pendingTerminalCloseTimers.set(sessionId, timer);
-}
-
-async function pollSessionFileBeforeClosing({
-  sessionId,
-  pollMs,
-  maxKeepAliveMs,
-  previousStat,
-  statusFailureCount,
-  startedAt,
-  generation,
-}: {
-  sessionId: string;
-  pollMs: number;
-  maxKeepAliveMs: number;
-  previousStat: SessionFileStat | null;
-  statusFailureCount: number;
-  startedAt: number;
-  generation: number;
-}) {
-  if (pendingTerminalCloseGenerations.get(sessionId) !== generation) {
-    return;
-  }
-
-  if (Date.now() - startedAt >= maxKeepAliveMs) {
-    closeScheduledTerminal(sessionId, generation);
-    return;
-  }
-
-  const [currentStat, terminalStatus] = await Promise.all([
-    statDesktopTerminalSessionFile(sessionId).catch(() => null),
-    getDesktopTerminalStatus(sessionId).catch(() => undefined),
-  ]);
-
-  if (pendingTerminalCloseGenerations.get(sessionId) !== generation) {
-    return;
-  }
-
-  const terminalStillRunning =
-    terminalStatus?.status === "starting" || terminalStatus?.status === "running";
-
-  if (terminalStatus === undefined) {
-    if (statusFailureCount + 1 >= MAX_TERMINAL_STATUS_FAILURES_BEFORE_CLOSE) {
-      closeScheduledTerminal(sessionId, generation);
-      return;
-    }
-
-    scheduleTerminalCloseWhenSessionFileIdle({
-      sessionId,
-      pollMs,
-      maxKeepAliveMs,
-      previousStat,
-      statusFailureCount: statusFailureCount + 1,
-      startedAt,
-      generation,
-    });
-    return;
-  }
-
-  if (!terminalStillRunning) {
-    closeScheduledTerminal(sessionId, generation);
-    return;
-  }
-
-  if (!currentStat) {
-    if (!previousStat) {
-      closeScheduledTerminal(sessionId, generation);
-      return;
-    }
-
-    scheduleTerminalCloseWhenSessionFileIdle({
-      sessionId,
-      pollMs,
-      maxKeepAliveMs,
-      previousStat: null,
-      statusFailureCount: 0,
-      startedAt,
-      generation,
-    });
-    return;
-  }
-
-  if (
-    previousStat &&
-    currentStat.mtimeMs === previousStat.mtimeMs &&
-    currentStat.size === previousStat.size
-  ) {
-    closeScheduledTerminal(sessionId, generation);
-    return;
-  }
-
-  scheduleTerminalCloseWhenSessionFileIdle({
-    sessionId,
-    pollMs,
-    maxKeepAliveMs,
-    previousStat: currentStat,
-    statusFailureCount: 0,
-    startedAt,
-    generation,
-  });
-}
-
-function scheduleTerminalCloseWhenSessionFileIdle({
-  sessionId,
-  pollMs,
-  maxKeepAliveMs,
-  previousStat,
-  statusFailureCount,
-  startedAt,
-  generation,
-}: {
-  sessionId: string;
-  pollMs: number;
-  maxKeepAliveMs: number;
-  previousStat: SessionFileStat | null;
-  statusFailureCount: number;
-  startedAt: number;
-  generation: number;
-}) {
-  const timer = setTimeout(() => {
-    void pollSessionFileBeforeClosing({
-      sessionId,
-      pollMs,
-      maxKeepAliveMs,
-      previousStat,
-      statusFailureCount,
-      startedAt,
-      generation,
-    });
-  }, pollMs);
-
-  pendingTerminalCloseTimers.set(sessionId, timer);
-}
-
-async function scheduleTerminalCloseAfterSessionFileIdle(
-  sessionId: string,
-  pollMs: number,
-  maxKeepAliveMs: number,
-) {
-  cancelScheduledTerminalClose(sessionId);
-  const generation = (pendingTerminalCloseGenerations.get(sessionId) ?? 0) + 1;
-  pendingTerminalCloseGenerations.set(sessionId, generation);
-  const startedAt = Date.now();
-  const initialStat = await statDesktopTerminalSessionFile(sessionId).catch(() => null);
-
-  if (pendingTerminalCloseGenerations.get(sessionId) !== generation) {
-    return;
-  }
-
-  scheduleTerminalCloseWhenSessionFileIdle({
-    sessionId,
-    pollMs,
-    maxKeepAliveMs,
-    previousStat: initialStat,
-    statusFailureCount: 0,
-    startedAt,
-    generation,
-  });
-}
-
-function writeSystemMessage(write: (data: string) => void, message: string) {
-  write(`\r\n[terminal] ${message}\r\n`);
-}
-
-function clearTerminal(write: (data: string) => void) {
-  write(CLEAR_TERMINAL_SEQUENCE);
-}
-
-function clampTerminalHistory(history: string) {
-  return history.length > MAX_FRONTEND_HISTORY_CHARS
-    ? history.slice(-TRIMMED_FRONTEND_HISTORY_CHARS)
-    : history;
-}
-
-function extractTerminalLinks(line: string): TerminalLinkMatch[] {
-  const matches = [...line.matchAll(TERMINAL_LINK_PATTERN)];
-  return matches.map((match) => ({
-    text: match[0],
-    start: match.index ?? 0,
-    end: (match.index ?? 0) + match[0].length,
-  }));
-}
-
-function getCaretPositionFromPoint(document: Document, clientX: number, clientY: number) {
-  const documentWithCaretApi = document as Document & {
-    caretPositionFromPoint?: (x: number, y: number) => { offsetNode: Node; offset: number } | null;
-    caretRangeFromPoint?: (x: number, y: number) => Range | null;
-  };
-
-  const caretPosition = documentWithCaretApi.caretPositionFromPoint?.(clientX, clientY);
-  if (caretPosition) {
-    return {
-      node: caretPosition.offsetNode,
-      offset: caretPosition.offset,
-    };
-  }
-
-  const caretRange = documentWithCaretApi.caretRangeFromPoint?.(clientX, clientY);
-  if (caretRange) {
-    return {
-      node: caretRange.startContainer,
-      offset: caretRange.startOffset,
-    };
-  }
-
-  return null;
-}
-
-function findTerminalRow(container: HTMLElement, node: Node | null) {
-  let current: Node | null = node;
-
-  while (current) {
-    if (current instanceof HTMLElement && current.classList.contains("term-row")) {
-      return container.contains(current) ? current : null;
-    }
-
-    current = current.parentNode;
-  }
-
-  return null;
-}
-
-function getRowTextOffset(row: HTMLElement, node: Node, offset: number) {
-  const range = row.ownerDocument.createRange();
-
-  try {
-    range.setStart(row, 0);
-    range.setEnd(node, offset);
-  } catch {
-    return null;
-  }
-
-  return range.toString().length;
-}
-
-function findTerminalLinkAtPoint(container: HTMLElement, clientX: number, clientY: number) {
-  const caret = getCaretPositionFromPoint(container.ownerDocument, clientX, clientY);
-  if (!caret) {
-    return null;
-  }
-
-  const row = findTerminalRow(container, caret.node);
-  if (!row) {
-    return null;
-  }
-
-  const rowTextOffset = getRowTextOffset(row, caret.node, caret.offset);
-  if (rowTextOffset === null) {
-    return null;
-  }
-
-  const lineText = row.textContent ?? "";
-  return (
-    extractTerminalLinks(lineText).find(
-      (match) => rowTextOffset >= match.start && rowTextOffset < match.end,
-    ) ?? null
-  );
-}
-
-function hasSelectionInside(container: HTMLElement) {
-  const selection = window.getSelection();
-  if (!selection || selection.isCollapsed || selection.toString().trim().length === 0) {
-    return false;
-  }
-
-  const anchorNode = selection.anchorNode;
-  const focusNode = selection.focusNode;
-
-  return Boolean(
-    (anchorNode && container.contains(anchorNode)) || (focusNode && container.contains(focusNode)),
-  );
-}
-
-function normalizeTerminalDimension(value: number, fallback: number) {
-  return Number.isFinite(value) && value > 0 ? Math.floor(value) : fallback;
-}
-
-function isUsableTerminalSize(cols: number, rows: number) {
-  return cols >= MIN_USABLE_TERMINAL_COLS && rows >= MIN_USABLE_TERMINAL_ROWS;
-}
-
-function measureTerminalSize(terminal: WTerm) {
-  const element = terminal.element;
-  const grid = element.querySelector<HTMLElement>(".term-grid");
-  if (!grid) {
-    return null;
-  }
-
-  const row = element.ownerDocument.createElement("div");
-  row.className = "term-row";
-  row.style.visibility = "hidden";
-  row.style.position = "absolute";
-
-  const probe = element.ownerDocument.createElement("span");
-  probe.textContent = "W";
-  row.appendChild(probe);
-  grid.appendChild(row);
-
-  const charWidth = probe.getBoundingClientRect().width;
-  const rowHeight = row.getBoundingClientRect().height;
-  row.remove();
-
-  if (charWidth <= 0 || rowHeight <= 0) {
-    return null;
-  }
-
-  const styles = getComputedStyle(element);
-  const paddingLeft = Number.parseFloat(styles.paddingLeft) || 0;
-  const paddingRight = Number.parseFloat(styles.paddingRight) || 0;
-  const paddingTop = Number.parseFloat(styles.paddingTop) || 0;
-  const paddingBottom = Number.parseFloat(styles.paddingBottom) || 0;
-  const contentWidth = element.clientWidth - paddingLeft - paddingRight;
-  const contentHeight = element.clientHeight - paddingTop - paddingBottom;
-
-  if (contentWidth <= 0 || contentHeight <= 0) {
-    return null;
-  }
-
-  return {
-    cols: Math.max(1, Math.floor(contentWidth / charWidth)),
-    rows: Math.max(1, Math.floor(contentHeight / rowHeight)),
-  };
-}
-
-function terminalWrapperStyle(backgroundCssVar: TerminalBackgroundCssVar): CSSProperties {
-  return {
-    "--terminal-surface": `var(${backgroundCssVar})`,
-  } as CSSProperties;
-}
-
-function terminalStyleVars(backgroundCssVar: TerminalBackgroundCssVar): CSSProperties {
-  return {
-    "--terminal-selection": "rgba(185, 191, 243, 0.18)",
-    "--term-bg": `var(${backgroundCssVar})`,
-    "--term-fg": "var(--text)",
-    "--term-cursor": "var(--accent)",
-    "--term-font-family": '"Liberation Mono", Consolas, Menlo, monospace',
-    "--term-font-size": "12px",
-    "--term-line-height": "1.2",
-    "--term-color-0": `var(${backgroundCssVar})`,
-    "--term-color-1": "#db7d84",
-    "--term-color-2": "var(--green)",
-    "--term-color-3": "var(--accent)",
-    "--term-color-4": "var(--accent)",
-    "--term-color-5": "var(--accent)",
-    "--term-color-6": "var(--muted)",
-    "--term-color-7": "var(--text)",
-    "--term-color-8": "var(--muted-2)",
-    "--term-color-9": "#ec979d",
-    "--term-color-10": "var(--green)",
-    "--term-color-11": "var(--text)",
-    "--term-color-12": "var(--text)",
-    "--term-color-13": "var(--text)",
-    "--term-color-14": "var(--text)",
-    "--term-color-15": "#f7f9ff",
-  } as CSSProperties;
-}
-
-function isTerminalElementNearBottom(element: HTMLElement) {
-  return (
-    element.scrollHeight - element.clientHeight - element.scrollTop <=
-    TERMINAL_STICKY_BOTTOM_THRESHOLD_PX
-  );
-}
 
 export function TerminalViewport({
   projectId,
